Gatewayless forwarding is not working with FIP enabled
In this case src_vn and dest_vn belong to different VNs and FIP is
attached to src VMI. Still it doesn't take FIP for forwarding due
to a new check added as part of CEM-12277. In case of CEM-18181,
there is an assymetric path for outgoing and incoming traffic. So,
incoming traffic comes on vhost0 and triggers a new flow creation
and traffic gets impacted.
Fix - Compare src_vn and dest_vn, if different then use FIP

closes-jira-bug: CEM-18181
Change-Id: I536d520a4311ea52ce174ea19bf460af649828c8