config/openstack: fix exception sent to neutron for port in-use
The L3PortInUse's exception sent to Neutron does not seem to exist
(checked version until Rocky). Hence it is not possible for Neutron to
handle it correctly to inform why the action was not performed.

This commit updates the exception to PortInUse which is a
comprhensible one for Neutron. It also adds missing test to assert
behavior when trying to delete port in-use.

Closes-Jira-Bug: CEM-16984
Signed-off-by: Sahid Orentino Ferdjaoui <sferdjaoui@juniper.net>
Change-Id: Ib492df9d6c9f18acfda4c80d69868e1ce33b575f