Trap BFD packets to agent for implementation of BFD priority queue
In scaled setups it was observed that the BFD packets to agent are handled
by the workqueue job in PktHandler task that has mutula dependency with
dB infra task. This is causing the stalling of the BFD packets and causing
the BFD to timeout on the established sessions. This is causing the interface
on which the BFD is run to flap and causing the other proto's to flaps as well.

To address this is,
1. In vrouter, trap the BFD packets that are destined to Agent.
2. In agent, use the BFD trap informantion from the vrouter to enqueue
and process BFD (keepalive) packets to a seperate queue.
The BFD control packets (states with Admindown, Down & Init) that has dependecy on Db
task are enqueued in separte queue.

BFD keepalive packets drop counter is added to Introspect for accounting
errors in the traped BFD trapped packets to Agent.

Unit tests are also added to verify the vrouter BFD Trap information.

Note BFD state-machine code in agent is unchanged.

patch changes-
1. Added BFD keepalive enqueue count in Agent Introspect
2. made inline function for checking the BFD keepalive pkts
   in ASIO task context
3. other cosmetic changes

Closes-jira-bug: CEM-16291
Depends-on: I15587a6cca4075f1384f9a9f3aac6a832f550c2c
Change-Id: Ie77131dbb26e0e84450105f17bcc136140ab7a33