Update patch set 26
Patch Set 26: Code-Review-1

I did not have time to review all of the logic behind but I would not recommend updating 50 resources that to add a new parameter for their returns that just for one which needs it. That also based on the fact you did not had to fix UTs which means that nothing are asserted.

Can we discuss and try to think of an other appraoch?

Patch-set: 26
Reviewer: Gerrit User 40 <40@c72cd27d-1743-4a55-84a4-cd13b8400c52>
Label: Code-Review=-1