Update flow and ksync_entry on EEXIST
In a corner case, two or more packets are trapped from vrouter for BGPaaS,
which will eventually cause the vrouter to receive an ADD request for the same
RF (from Controller) and then return EEXIST.
For this case we decide to delete that old flow, but we need to make sure that
the ksync_entry flow handle is also updated.

Ported the changes of review (https://review.opencontrail.org/c/Juniper/contrail-controller/+/62957) for a hot fix branch.

Closes-Jira-bug: CEM-18166

Signed-off-by: NagarajM R <mrn@juniper.net>
Change-Id: Id74f844b6229f450d2d64126fffdb4f63cc7d6b2