Delete FMG nh list entries of previous VRFs when lable is reused.
when there is a change in label for multicast tree, old label
delete is not happening sometimes, so vrf entry is present
in old fmg label and holds reference for VRF and there is no way
to recover if it happens.
made changes to to delete existing VRF entries from fmg NH list of
label when it is reused for multicast tree for different VRF.

This change is requested by a customer for R1912.L3 (CEM-20098).

Ported the changes of review (https://review.opencontrail.org/c/Juniper/contrail-controller/+/62935) for a hot fix branch.
Closes-Jira-Bug: CEM-19981

Signed-off-by: NagarajM R <mrn@juniper.net>
Change-Id: I5960dd7b034d8aa63f0e0548741c24a0b8fa955c