Fix for CEM-20974
Problem:
1. Keystone binds port 5000 in public network and 35357
in internal network.
2. Default auth_url populated in contrail-keysone-auth.conf
uses port 35357. It has connectivity to keystone only via public network
and hence config-api is not able to reach keystone and sync projects.

Fix:
This fix adds an endpoint_override argument in Keystone (v2 and v3)
client. When endpoint_override is passed to the client call, the client
need not discover the list of endpoints again and can simply use the
_auth_url provided.

partial-jira-bug: CEM-20974

Change-Id: I3a6c4a44857a5c73b14bfa6d4a8718f3edb0a047
Signed-off-by: Sayali Mane <smane@juniper.net>