NIMS - BER500 - Live - vRouter Agent Crash with VrfDeleteTimeout
Problem: In a test of continuous VM restarted, Vrf delete is not successful even after the timeout period of 150 minutes. Thus it asserts after the time expiry.

Root cause: When a last VM is restarted from a network, it triggers the vrf deletion. VRF and its related objects are marked for delete. In certain cases, its related object(interface) is being reused but VRF is still delete marked. This makes the VRF to get timedout as its interface is reused and stay active. There are 2 cases this can occur.
	1. On VM up, VMI add notification has reached to operDB. OperDB does a lookup and it finds the entry which was is delete marked. OperDB reused the interface object by clearing the delete marked flag.
	2. VMI has got deleted but immediately OperDB receives a VMI add notification and created the VMI before VRF is deleted. Eventually it refers to the same VRF and it results in the same timeout assert.

Solution: On the above 2 cases, reuse the VRF by clearing the delete flag and reset the timeout timer.

NOTE: current review changes has dependency on https://review.opencontrail.org/c/Juniper/contrail-common/+/63933. Please approve it first to avoid the compilation error after the merge.

Closes-Jira-Bug: CEM-24751

Signed-off-by: praveen kaliaperumal <pkaliaperuma@juniper.net>
Depends-On: I5b16af58882ba58351416e145dcda2beffd4559c
Change-Id: I24c04d62a4bb0f7f379a299a029647eefadbd053