Do not update response for default dns
For default dns method, DNS resolution for the VMs is done based
on the name server configuration in the server infrastructure.
Also query done with host=subnet .1 and .2 ip is also proxied to
name server in underlay network. Dns response message can come
with a different offest , since no update for offset is done for
this mode, do not change the offset in dns response message, just
forward response to vm.

Change-Id: I5ec254f663d9b04af570afc8bcbca155018cd737